refactor: *Options.ts moved to options/ from util/, fixes to the allowedmentions typings
