+++ packages/tiscord/src/index.ts
@@ -1,22 +1,22 @@
-export { BaseThreadOptions, ThreadData, ThreadOptions } from './util/ThreadOptions';
-export { GuildEditOptions, GuildEditOptionsType } from './util/GuildEditOptions';
-export { InviteData, InviteOptions } from './util/InviteOptions';
-export { MemberOptions, RawMemberOptions } from './util/MemberOptions';
-export { MessageOptions, RawMessageOptions } from './util/MessageOptions';
+export { BaseThreadOptions, ThreadData, ThreadOptions } from './options/ThreadOptions';
+export { GuildEditOptions, GuildEditOptionsType } from './options/GuildEditOptions';
+export { InviteData, InviteOptions } from './options/InviteOptions';
+export { MemberOptions, RawMemberOptions } from './options/MemberOptions';
+export { MessageOptions, RawMessageOptions } from './options/MessageOptions';
 export { MessageTypes, SystemMessageTypes } from './util/MessageTypes';
 
 export { APIError } from './errors/APIError';
 export { Application } from './structures/application/Application';
 export { MessageAttachment } from './structures/message/MessageAttachment';
 export { BitField } from './util/BitField';
-export { CDNOptions } from './util/CDNOptions';
+export { CDNOptions } from './options/CDNOptions';
 export { Cache } from './cache/Cache';
-export { CacheOptions } from './util/CacheOptions';
+export { CacheOptions } from './options/CacheOptions';
 export { Channel } from './structures/channels/Channel';
 export { ChannelManager } from './managers/ChannelManager';
-export { ChannelOptions } from './util/ChannelOptions';
+export { ChannelOptions } from './options/ChannelOptions';
 export { Client } from './client/Client';
-export { ClientOptions } from './util/ClientOptions';
+export { ClientOptions } from './options/ClientOptions';
 export { DMChannel } from './structures/channels/DMChannel';
 export { Emoji } from './structures/guild/Emoji';
 export { EventManager } from './managers/EventManager';
@@ -29,7 +29,7 @@ export { GuildChannel } from './structures/channels/GuildChannel';
 export { GuildManager } from './managers/GuildManager';
 export { Integration } from './structures/guild/Integration';
 export { Interaction } from './structures/interactions/Interaction';
-export { InteractionOptions } from './util/InteractionOptions';
+export { InteractionOptions } from './options/InteractionOptions';
 export { Invite } from './structures/guild/Invite';
 export { Member } from './structures/guild/Member';
 export { MemberEdit } from './util/MemberEdit';
@@ -39,7 +39,7 @@ export { MessageCache } from './cache/MessageCache';
 export { MessageManager } from './managers/MessageManager';
 export { Permissions } from './util/Permissions';
 export { Role } from './structures/guild/Role';
-export { RoleOptions } from './util/RoleOptions';
+export { RoleOptions } from './options/RoleOptions';
 export { RoleManager } from './managers/RoleManager';
 export { SelectMenuOption } from './structures/interactions/SelectMenuOption';
 export { Sticker } from './structures/guild/Sticker';
