+++ packages/tiscord/src/util/AllowedMentions.ts
@@ -1,17 +1,19 @@
+import { AllowedMentionsTypes } from 'discord-api-types/v10';
+
 export class AllowedMentions {
-    parse: ('roles' | 'users' | 'everyone')[];
-    replied_uses: boolean;
-    roles: string[];
-    users: string[];
+    parse?: AllowedMentionsTypes[];
+    replied_user?: boolean;
+    roles?: string[];
+    users?: string[];
     constructor(mentions: RawMentions) {
         if (mentions.parse) this.parse = mentions.parse;
         if (mentions.roles) this.roles = mentions.roles;
         if (mentions.users) this.users = mentions.users;
-        if (mentions.repliedUser) this.replied_uses = mentions.repliedUser;
+        if (mentions.repliedUser) this.replied_user = mentions.repliedUser;
     }
 }
 export interface RawMentions {
-    parse?: ('users' | 'roles' | 'everyone')[];
+    parse?: AllowedMentionsTypes[];
     roles?: string[];
     users?: string[];
     repliedUser?: boolean;
